requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asttokens=2.4.1=pyhd8ed1ab_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.0=mkl
bottleneck=1.3.5=py39h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2024.7.2=h06a4308_0
certifi=2024.7.4=py39h06a4308_0
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd3eb1b0_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.0.5=py39hdb19cb5_0
cryptography=37.0.1=py39h9ce1e76_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.127=0
cuda-runtime=12.1.0=0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.12.0=py39h5eee18b_0
dask-core=2022.7.0=py39h06a4308_0
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py39h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
einops=0.4.1=pypi_0
et_xmlfile=1.1.0=py39h06a4308_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.8.0=pypi_0
fontconfig=2.14.1=hef1e5e3_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
fsspec=2022.11.0=py39h06a4308_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
huggingface-hub=0.11.0=pypi_0
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
imageio=2.19.3=py39h06a4308_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.26.0=pyhf8b6a83_0
ipython=8.17.2=pyh41d4057_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=py39h06a4308_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=4.12.0=py39hf3d152e_0
kiwisolver=1.4.4=py39h6a678d5_0
krb5=1.19.4=h568e23c_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm10=10.0.1=hbcb73fb_5
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h2818925_1
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
llvm-openmp=14.0.6=h9e868ea_0
locket=1.0.0=py39h06a4308_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.1.3=py39h5eee18b_0
matplotlib=3.6.2=py39h06a4308_0
matplotlib-base=3.6.2=py39h945d387_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.8=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py39h06a4308_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.3=py39h807cd23_0
numpy=1.22.3=py39he7a7128_0
numpy-base=1.22.3=py39hf524024_0
openh264=2.1.1=h4ff587b_0
openpyxl=3.0.10=py39h5eee18b_0
openssl=1.1.1w=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.3=py39h6a678d5_0
parso=0.8.3=pyhd8ed1ab_0
partd=1.2.0=pyhd3eb1b0_1
pcre=8.45=h295c915_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.0.1=py39h22f2fdc_0
pip=21.2.4=py39h06a4308_0
ply=3.11=py39h06a4308_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
psutil=5.9.0=py39h5eee18b_0
ptflops=0.6.9=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.16.1=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39h06a4308_0
pyqt=5.15.7=py39h6a678d5_1
pyqt5-sip=12.11.0=py39h6a678d5_1
pysocks=1.7.1=py39h06a4308_0
python=3.9.12=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=2.4.0=py3.9_cuda12.1_cudnn9.1.0_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2022.1=py39h06a4308_0
pywavelets=1.4.1=py39h5eee18b_0
pyyaml=6.0=py39h5eee18b_1
pyzmq=25.1.0=py39h6a678d5_0
qt-main=5.15.2=h327a75a_6
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.1.2=h7f8727e_1
requests=2.28.0=py39h06a4308_0
safetensors=0.4.2=pypi_0
scikit-image=0.16.2=py39ha9443f7_0
scikit-learn=1.0.2=py39h51133e4_1
scipy=1.7.3=py39hc147768_0
seaborn=0.13.2=py39h06a4308_0
setuptools=61.2.0=py39h06a4308_0
sip=6.6.2=py39h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.5=hc218d9a_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=py39h06a4308_0
threadpoolctl=2.2.0=pyh0d69192_0
timm=1.0.11=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=py39h06a4308_0
torchaudio=2.4.0=py39_cu121
torchtriton=3.0.0=py39
torchvision=0.19.0=py39_cu121
tornado=6.2=py39h5eee18b_0
tqdm=4.64.1=pypi_0
traitlets=5.13.0=pyhd8ed1ab_0
typing-extensions=4.12.2=pypi_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39h06a4308_0
wcwidth=0.2.9=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
